+++ src/components/FavoriteArtist/index.jsx
