+++ src/components/NewReleaseSlider/index.js
