+++ src/components/PlaylistSlider/index.js
