+++ src/pages/index.jsx
@@ -3,8 +3,8 @@ import Head from "next/head";
 import React from "react";
 import styles from "../../styles/App.module.scss";
 import useTrans from "../hooks/useStrans";
-import LayOut from "../components/layout";
-import Home from "../components/home";
+import Layout from "../components/Layout";
+import Home from "../components/Home";
 
 const App = () => {
   const trans = useTrans();
@@ -18,9 +18,9 @@ const App = () => {
           href="https://static-zmp3.zmdcdn.me/skins/zmp3-v5.2/images/icon_zing_mp3_60.png"
         />
       </Head>
-      <LayOut trans={trans}>
+      <Layout trans={trans}>
         <Home />
-      </LayOut>
+      </Layout>
     </div>
   );
 };
